Changes so it works for lukas
In plotteprofiler:
Just added directories into this file that correspond to directories on my computer so it's quicker to work on in the future. These can be commented out later.

In Sammenlikne:
Added legend. Spent time figuring out how to change the spacing in the figure window to fit everything. This may need to be adjusted when we have even more data in the legend. Or maybe just change font size again.